+++ src/main/kotlin/com/github/sam0delkin/intellijpsa/action/MultipleFileTemplateAction.kt
@@ -1,3 +1,5 @@
+@file:Suppress("ktlint:standard:no-wildcard-imports")
+
 package com.github.sam0delkin.intellijpsa.action
 
 import com.github.sam0delkin.intellijpsa.services.CompletionService
@@ -5,11 +7,13 @@ import com.github.sam0delkin.intellijpsa.settings.TemplateFormField
 import com.github.sam0delkin.intellijpsa.settings.TemplateFormFieldType
 import com.github.sam0delkin.intellijpsa.ui.components.JTextFieldCollection
 import com.intellij.codeInsight.lookup.LookupElementBuilder
+import com.intellij.icons.AllIcons
 import com.intellij.ide.IdeView
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.DataContext
 import com.intellij.openapi.actionSystem.LangDataKeys
+import com.intellij.openapi.actionSystem.impl.ActionButton
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.application.ex.ApplicationUtil
 import com.intellij.openapi.components.service
@@ -17,44 +21,47 @@ import com.intellij.openapi.editor.EditorFactory
 import com.intellij.openapi.editor.event.DocumentEvent
 import com.intellij.openapi.editor.event.DocumentListener
 import com.intellij.openapi.fileEditor.OpenFileDescriptor
+import com.intellij.openapi.fileTypes.FileType
 import com.intellij.openapi.fileTypes.FileTypeManager
 import com.intellij.openapi.fileTypes.FileTypes
 import com.intellij.openapi.progress.EmptyProgressIndicator
 import com.intellij.openapi.progress.ProcessCanceledException
+import com.intellij.openapi.project.DumbAwareAction
 import com.intellij.openapi.project.guessProjectDir
 import com.intellij.openapi.ui.ComboBox
 import com.intellij.openapi.ui.DialogBuilder
 import com.intellij.openapi.util.NlsActions
 import com.intellij.openapi.util.text.StringUtil
 import com.intellij.openapi.vfs.VfsUtil
+import com.intellij.openapi.vfs.VirtualFileManager
 import com.intellij.psi.PsiDirectory
-import com.intellij.psi.PsiFile
-import com.intellij.psi.PsiFileFactory
-import com.intellij.psi.impl.file.PsiDirectoryFactory
 import com.intellij.ui.AnimatedIcon
 import com.intellij.ui.EditorTextField
 import com.intellij.ui.HorizontalScrollBarEditorCustomization
 import com.intellij.ui.components.JBCheckBox
+import com.intellij.ui.components.JBLabel
+import com.intellij.ui.components.JBTabbedPane
 import com.intellij.ui.components.JBTextField
-import com.intellij.ui.dsl.builder.Cell
-import com.intellij.ui.dsl.builder.panel
+import com.intellij.ui.dsl.builder.*
 import com.intellij.util.textCompletion.TextCompletionValueDescriptor
 import com.intellij.util.textCompletion.TextFieldWithCompletion
 import com.intellij.util.textCompletion.ValuesCompletionProvider
 import kotlinx.serialization.json.JsonArray
 import kotlinx.serialization.json.JsonObject
+import kotlinx.serialization.json.jsonArray
 import kotlinx.serialization.json.jsonPrimitive
 import java.awt.Dimension
+import java.awt.FlowLayout
+import java.awt.Point
+import java.nio.charset.StandardCharsets
+import java.nio.file.Files
+import java.nio.file.Path
+import java.nio.file.Paths
 import java.util.Timer
 import java.util.TimerTask
-import javax.swing.Icon
-import javax.swing.JCheckBox
-import javax.swing.JComponent
-import javax.swing.JLabel
-import javax.swing.SwingUtilities
-import kotlin.collections.ArrayList
+import javax.swing.*
 
-class SingleFileTemplateAction(
+class MultipleFileTemplateAction(
     text: @NlsActions.ActionText String?,
     description: @NlsActions.ActionDescription String?,
     icon: Icon?,
@@ -62,9 +69,12 @@ class SingleFileTemplateAction(
     var templateName: String? = null
     private var changed: Boolean = true
     private var timer: Timer? = null
-    private var fileNameField: Cell<JLabel>? = null
-    private var previewTextField: Cell<EditorTextField>? = null
+    private var filePathFields: ArrayList<JLabel> = ArrayList()
+    private var filePaths: ArrayList<String> = ArrayList()
+    private var tabbedPane: JBTabbedPane? = null
+    private var previewTextFields: ArrayList<EditorTextField> = ArrayList()
     private var loadingIcon: Cell<JLabel>? = null
+    private var errorIcon: Cell<ActionButton>? = null
     private var navigateToFile: Cell<JCheckBox>? = null
     private val formFields: HashMap<String, Cell<JComponent>> = HashMap()
     private val richTextEditorValues = HashMap<String, MutableList<String>>()
@@ -78,7 +88,7 @@ class SingleFileTemplateAction(
         }
 
         val settings = completionService.getSettings()
-        val template = settings.singleFileCodeTemplates?.find { it.name == templateName }
+        val template = settings.multipleFileCodeTemplates?.find { it.name == templateName }
 
         if (null === template || null === template.formFields) {
             return
@@ -107,29 +117,41 @@ class SingleFileTemplateAction(
                     e.project!!,
                     directoryPath,
                     template.name!!,
-                    "single_file",
+                    "multiple_file",
                     originatorFieldName,
                     formFieldData,
                 )
 
             if (null === templateData) {
+                loadingIcon?.visible(false)
+                errorIcon?.visible(true)
+
                 return
             }
 
             SwingUtilities.invokeLater {
                 ApplicationManager.getApplication().runWriteAction {
-                    if (!templateData.containsKey("file_name")) {
+                    if (!templateData.containsKey("file_names")) {
                         return@runWriteAction
                     }
+                    val fileTypes = ArrayList<FileType>()
 
-                    val newFileType =
-                        FileTypeManager
-                            .getInstance()
-                            .getFileTypeByFileName(templateData.get("file_name")!!.jsonPrimitive.content)
-                    if (templateData.containsKey("file_name")) {
-                        if (null !== fileNameField) {
-                            fileNameField!!.component.text = templateData.get("file_name")!!.jsonPrimitive.content
-                        }
+                    filePaths.clear()
+                    for (key in 0 until templateData.get("file_names")!!.jsonArray.size) {
+                        val filePath =
+                            templateData
+                                .get("file_names")!!
+                                .jsonArray[key]
+                                .jsonPrimitive.content
+                        val fileName = filePath.split('/').last()
+                        val newFileType =
+                            FileTypeManager
+                                .getInstance()
+                                .getFileTypeByFileName(fileName)
+                        fileTypes.add(newFileType)
+                        this.tabbedPane!!.setTitleAt(key, fileName)
+                        filePathFields[key].text = "File Path: " + filePath
+                        filePaths.add(filePath)
                     }
 
                     if (templateData.containsKey("form_fields")) {
@@ -173,14 +195,15 @@ class SingleFileTemplateAction(
                         }
                     }
 
-                    if (templateData.containsKey("content")) {
-                        if (null !== previewTextField) {
-                            if (newFileType !== previewTextField!!.component.fileType) {
-                                previewTextField!!.component.fileType = newFileType
+                    if (templateData.containsKey("contents")) {
+                        for (index in 0 until templateData.get("contents")!!.jsonArray.size) {
+                            val previewTextField = previewTextFields[index]
+                            val newFileType = fileTypes[index]
+                            if (newFileType !== previewTextField.fileType) {
+                                previewTextField.fileType = newFileType
                             }
-                            previewTextField!!.component.document.setText(
-                                StringUtil.convertLineSeparators(templateData.get("content")!!.jsonPrimitive.content),
-                            )
+                            val contents = templateData.get("contents")!!.jsonArray[index]
+                            previewTextField.document.setText(contents.jsonPrimitive.content)
                         }
                     }
                 }
@@ -196,6 +219,7 @@ class SingleFileTemplateAction(
         ) {
             changed = true
             loadingIcon?.visible(true)
+            errorIcon?.visible(false)
             if (field !== null && value !== null) {
                 formFieldData[field.name!!] = value
             }
@@ -212,9 +236,12 @@ class SingleFileTemplateAction(
                     object : TimerTask() {
                         override fun run() {
                             try {
-                                ApplicationUtil.runWithCheckCanceled({
-                                    updateData(field?.name)
-                                }, indicator)
+                                ApplicationUtil.runWithCheckCanceled(
+                                    {
+                                        updateData(field?.name)
+                                    },
+                                    indicator,
+                                )
                             } catch (_: ProcessCanceledException) {
                             }
                         }
@@ -291,51 +318,82 @@ class SingleFileTemplateAction(
                                     },
                                 )
                                 richTextEditorValues[field.name!!] = values
-                                val cell = cell(richText)
-                                formFields[field.name!!] = cell
+                                cell(richText)
                             }
-
-                            if (null !== field.focused && field.focused!!) {
-                                SwingUtilities.invokeLater {
-                                    formFields[field.name!!]!!.component.requestFocus()
-                                }
-                            }
-                            formFields[field.name!!]!!.component.minimumSize = Dimension(600, 30)
-                            formFields[field.name!!]!!.component.preferredSize = Dimension(600, 30)
                         }
-                    }
-                    row("File Name") {
-                        fileNameField = label("")
                     }
                     row {
                         label("Preview")
                         loadingIcon = icon(AnimatedIcon.Default()).visible(false)
+                        @Suppress("DialogTitleCapitalization")
+                        val action =
+                            object : DumbAwareAction(
+                                "Get List of supported languages by your IDE",
+                                "",
+                                AllIcons.General.ErrorDialog,
+                            ) {
+                                override fun actionPerformed(e: AnActionEvent) {
+                                    val tooltip =
+                                        com.intellij.ui
+                                            .GotItTooltip(
+                                                "PSA",
+                                                completionService.lastResultMessage,
+                                            ).withIcon(AllIcons.General.ErrorDialog)
+                                            .withButtonLabel("OK")
+                                    tooltip.createAndShow(errorIcon!!.component) { c, _ ->
+                                        Point(
+                                            c.width,
+                                            c.height / 2,
+                                        )
+                                    }
+                                }
+                            }
+                        errorIcon = actionButton(action).visible(false)
                     }
                     row { }
                     row {
-                        val previewTextFieldComponent =
-                            EditorTextField(
-                                EditorFactory.getInstance().createDocument(
-                                    StringUtil.convertLineSeparators(""),
-                                ),
-                                e.project,
-                                FileTypes.PLAIN_TEXT,
-                                true,
-                                false,
-                            )
-                        previewTextFieldComponent.addSettingsProvider { editor ->
-                            run {
-                                editor.settings.isLineNumbersShown = true
-                                editor.setVerticalScrollbarVisible(true)
-                                HorizontalScrollBarEditorCustomization.ENABLED.customize(editor)
+                        panel {
+                            row {
+                                tabbedPane = JBTabbedPane()
+
+                                for (i in 0 until template.fileCount!!) {
+                                    val panel = JPanel(FlowLayout(FlowLayout.LEFT))
+                                    panel.preferredSize = Dimension(800, 600)
+                                    panel.minimumSize = Dimension(800, 200)
+                                    panel.maximumSize = Dimension(-1, 600)
+                                    val previewTextFieldComponent =
+                                        EditorTextField(
+                                            EditorFactory.getInstance().createDocument(
+                                                StringUtil.convertLineSeparators(""),
+                                            ),
+                                            e.project,
+                                            FileTypes.PLAIN_TEXT,
+                                            true,
+                                            false,
+                                        )
+                                    previewTextFieldComponent.addSettingsProvider { editor ->
+                                        run {
+                                            editor.settings.isLineNumbersShown = true
+                                            editor.setVerticalScrollbarVisible(true)
+                                            HorizontalScrollBarEditorCustomization.ENABLED.customize(editor)
+                                        }
+                                    }
+                                    previewTextFieldComponent.preferredSize = Dimension(800, 600)
+                                    previewTextFieldComponent.minimumSize = Dimension(800, 200)
+                                    previewTextFieldComponent.maximumSize = Dimension(-1, 600)
+                                    previewTextFields.add(previewTextFieldComponent)
+                                    filePathFields.add(JBLabel("File Path: "))
+                                    panel.add(filePathFields[i])
+                                    panel.add(previewTextFieldComponent)
+                                    panel.isVisible = true
+                                    tabbedPane!!.addTab("File 1", AllIcons.FileTypes.Any_type, panel)
+                                }
+
+                                cell(tabbedPane!!)
                             }
                         }
-                        previewTextFieldComponent.preferredSize = Dimension(800, 600)
-                        previewTextFieldComponent.minimumSize = Dimension(800, 200)
-                        previewTextFieldComponent.maximumSize = Dimension(-1, 600)
-                        previewTextField = cell(previewTextFieldComponent)
                     }.resizableRow()
-                    row("Navigate to file after creation") {
+                    row("Navigate to files after creation") {
                         navigateToFile = checkBox("")
                         navigateToFile!!.component.isSelected = true
                     }
@@ -345,22 +403,21 @@ class SingleFileTemplateAction(
         dialogBuilder.setTitle("Create " + template.title)
         if (dialogBuilder.showAndGet()) {
             ApplicationManager.getApplication().runWriteAction {
-                val fileFromText: PsiFile =
-                    PsiFileFactory
-                        .getInstance(e.project)
-                        .createFileFromText(
-                            fileNameField!!.component.text,
-                            previewTextField!!.component.fileType,
-                            previewTextField!!.component.document.text,
-                        )
-                val el =
-                    PsiDirectoryFactory
-                        .getInstance(e.project)
-                        .createDirectory(directories[0].virtualFile)
-                        .add(fileFromText)
+                var index = 0
+                filePaths.forEach {
+                    val previewTextField = previewTextFields[index]
+                    val fullFileName = e.project!!.guessProjectDir()!!.path + "/" + it
+                    val fileNameParts = fullFileName.split("/")
+                    val filePath = fileNameParts.dropLast(1).joinToString("/")
+                    Files.createDirectories(Paths.get(filePath))
+                    val file: Path = Paths.get(fullFileName)
+                    Files.write(file, previewTextField.document.text.split("\n"), StandardCharsets.UTF_8)
+                    val virtualFile = VirtualFileManager.getInstance().refreshAndFindFileByUrl("file://" + fullFileName)
 
-                if (navigateToFile!!.component.isSelected) {
-                    OpenFileDescriptor(e.project!!, el.containingFile.virtualFile, 0).navigate(true)
+                    if (navigateToFile!!.component.isSelected && virtualFile !== null) {
+                        OpenFileDescriptor(e.project!!, virtualFile, 0).navigate(true)
+                    }
+                    index++
                 }
             }
         }