+++ src/main/kotlin/com/github/sam0delkin/intellijpsa/listener/PsaOpenedFileListener.kt
@@ -1,6 +1,7 @@
 package com.github.sam0delkin.intellijpsa.listener
 
 import com.github.sam0delkin.intellijpsa.index.PsaIndex
+import com.github.sam0delkin.intellijpsa.settings.Settings
 import com.intellij.openapi.components.service
 import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.fileEditor.FileEditorManagerEvent
@@ -13,7 +14,14 @@ class PsaOpenedFileListener : FileEditorManagerListener {
         source: FileEditorManager,
         file: VirtualFile,
     ) {
+        val settings = source.project.service<Settings>()
+
+        if (!settings.indexingEnabled) {
+            return
+        }
+
         val psiFile = PsiManager.getInstance(source.project).findFile(file)
+
         if (null === psiFile) {
             return
         }
@@ -24,6 +32,12 @@ class PsaOpenedFileListener : FileEditorManagerListener {
 
     override fun selectionChanged(event: FileEditorManagerEvent) {
         if (null === event.newFile) {
+            return
+        }
+
+        val settings = event.manager.project.service<Settings>()
+
+        if (!settings.indexingEnabled) {
             return
         }